+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 * Fixed VersionMismatch parsing error [#2](https://github.com/contentful/contentful-management.py/issues/2)
 
+### Changed
+
+* Reuse result from `#publish` for reloading [#1](https://github.com/contentful/contentful-management.py/issues/1)
+
 ## v1.0.2
 
 * Add `snapshot` property to Snapshot API.