+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 * Added check for undefined fields from the WebApp against presence in the Content Type
 
+### Changed
+
+* Reuse result from `#publish` for reloading [#1](https://github.com/contentful/contentful-management.py/issues/1)
+
 ## v1.0.2
 
 * Add `snapshot` property to Snapshot API.