+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 * Added check for undefined fields from the WebApp against presence in the Content Type
 
+### Fixed
+
+* Fixed VersionMismatch parsing error [#2](https://github.com/contentful/contentful-management.py/issues/2)
+
 ### Changed
 
 * Reuse result from `#publish` for reloading [#1](https://github.com/contentful/contentful-management.py/issues/1)